feat(sirduck): Add simple constructor and create two tests
Create SirDuck.kt and the main class of the bot with a constructor for
the JDA interface as example for showing the two simple unit tests,
which mock the JDA interface, one with a successful connection to the
gateway, one with a shutdown.
It is just for demonstration, DI will be integrated, it won't stay like
that.